Reorder ntp verification tasks
The NTP-related checks were implemented to help Admins debug common
install-time misconfigurations of DNS servers (which would break package
fetching via apt). We've since moved the kernel install tasks to precede
the ntp checks, so they're not really serving their purpose.

On Trusty, ntp and ntpdate are present in the default configuration, and
therefore available for use. The same is not true of Xenial, however, so
let's first ensure that ntp and ntpdate are installed before referencing
them in commands. This is a bit of a catch-22, since any DNS failures
would cause the install tasks to fail. Still, though, it's worth
checking since these verification tasks may uncover errors related
strictly to NTP but not to DNS.