Ensure aptitude is installed
Aptitude is required for the subsequent package inspection tasks,
preventing tor breakage during playbook runs, as well as for using the
upgrade parameter in the Ansible 'apt' module. Under Trusty it's
installed by default, but not under Xenial. So let's make sure it's
present before referencing it in the config.